backend/requirements.txt
fastapi[all]==0.99.0
uvicorn[standard]==0.22.0
python-nmap==0.7.1
scapy==2.7.0
geoip2==5.2.0
requests==2.31.0
python-dotenv==1.0.0
